Move Usage interface implementations into separate files
This allows them to be ignored by goverter during build. It doesn't
fully work, but it's a step in the right direction.
