+++ apis/protection/v1beta1/usage_types.go
@@ -1,5 +1,3 @@
-//go:build !goverter
-
 /*
 Copyright 2025 The Crossplane Authors.
 
@@ -22,8 +20,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
-
-	"github.com/crossplane/crossplane/internal/protection"
 )
 
 // ResourceRef is a reference to a resource. It's suitable for cluster
@@ -167,71 +163,6 @@ type Usage struct {
 	metav1.ObjectMeta `json:"metadata,omitempty"`
 	Spec              UsageSpec   `json:"spec"`
 	Status            UsageStatus `json:"status,omitempty"`
-}
-
-// GetUserOf gets the resource this Usage indicates a use of.
-func (u *Usage) GetUserOf() protection.Resource {
-	conv := GeneratedNamespacedResourceConverter{}
-	return conv.ToInternal(u.Spec.Of)
-}
-
-// SetUserOf sets the resource this Usage indicates a use of.
-func (u *Usage) SetUserOf(r protection.Resource) {
-	conv := GeneratedNamespacedResourceConverter{}
-	u.Spec.Of = conv.FromInternal(r)
-}
-
-// GetUsedBy gets the resource this Usage indicates a use by.
-func (u *Usage) GetUsedBy() *protection.Resource {
-	if u.Spec.By == nil {
-		return nil
-	}
-	conv := GeneratedResourceConverter{}
-	out := conv.ToInternal(*u.Spec.By)
-	return &out
-}
-
-// SetUsedBy sets the resource this Usage indicates a use by.
-func (u *Usage) SetUsedBy(r *protection.Resource) {
-	if r == nil {
-		u.Spec.By = nil
-		return
-	}
-	conv := GeneratedResourceConverter{}
-	out := conv.FromInternal(*r)
-	u.Spec.By = &out
-}
-
-// GetReason gets the reason this Usage exists.
-func (u *Usage) GetReason() *string {
-	return u.Spec.Reason
-}
-
-// SetReason sets the reason this Usage exists.
-func (u *Usage) SetReason(reason *string) {
-	u.Spec.Reason = reason
-}
-
-// GetReplayDeletion gets a boolean that indicates whether deletion of the used
-// resource will be replayed when this Usage is deleted.
-func (u *Usage) GetReplayDeletion() *bool {
-	return u.Spec.ReplayDeletion
-}
-
-// SetReplayDeletion specifies whether deletion of the used resource will be
-// replayed when this Usage is deleted.
-func (u *Usage) SetReplayDeletion(replay *bool) {
-	u.Spec.ReplayDeletion = replay
-}
-
-// GetCondition of this Usage.
-func (u *Usage) GetCondition(ct xpv1.ConditionType) xpv1.Condition {
-	return u.Status.GetCondition(ct)
-}
-
-// SetConditions of this Usage.
-func (u *Usage) SetConditions(c ...xpv1.Condition) {
-	u.Status.SetConditions(c...)
 }
 
 // +kubebuilder:object:root=true
